Moved time functions from pkg stringz to pkg timez
