+++ cli/output/markdownw/markdownw.go
@@ -11,6 +11,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/neilotoole/sq/libsq/core/timez"
+
 	"github.com/neilotoole/sq/libsq/core/kind"
 	"github.com/neilotoole/sq/libsq/core/sqlz"
 	"github.com/neilotoole/sq/libsq/core/stringz"
@@ -91,11 +93,11 @@ func (w *RecordWriter) writeRecord(rec sqlz.Record) error {
 		case *time.Time:
 			switch w.recMeta[i].Kind() { //nolint:exhaustive
 			default:
-				s = val.Format(stringz.DatetimeFormat)
+				s = val.Format(timez.DatetimeFormat)
 			case kind.Time:
-				s = val.Format(stringz.TimeFormat)
+				s = val.Format(timez.TimeFormat)
 			case kind.Date:
-				s = val.Format(stringz.DateFormat)
+				s = val.Format(timez.DateFormat)
 			}
 		}
