+++ cli/output/xlsxw/xlsxw.go
@@ -5,10 +5,10 @@ import (
 	"io"
 	"time"
 
-	"github.com/neilotoole/sq/libsq/core/kind"
-	"github.com/neilotoole/sq/libsq/core/stringz"
+	"github.com/neilotoole/sq/libsq/core/timez"
 
 	"github.com/neilotoole/sq/cli/output"
+	"github.com/neilotoole/sq/libsq/core/kind"
 
 	"github.com/tealeg/xlsx/v2"
 
@@ -97,7 +97,7 @@ func (w *recordWriter) WriteRecords(recs []sqlz.Record) error {
 					// TODO: Maybe there's a way of setting a specific
 					//  time (as opposed to date or datetime) value, but
 					//  for now we just use a string.
-					cell.SetValue(val.Format(stringz.TimeFormat))
+					cell.SetValue(val.Format(timez.TimeFormat))
 				}
 			default:
 				// should never happen
