+++ cli/output/xmlw/xmlw.go
@@ -10,6 +10,8 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/neilotoole/sq/libsq/core/timez"
+
 	"github.com/fatih/color"
 
 	"github.com/neilotoole/sq/libsq/core/kind"
@@ -198,11 +200,11 @@ func (w *recordWriter) writeRecord(rec sqlz.Record) error {
 		case *time.Time:
 			switch w.recMeta[i].Kind() { //nolint:exhaustive
 			default:
-				w.fieldPrintFns[i](w.outBuf, val.Format(stringz.DatetimeFormat))
+				w.fieldPrintFns[i](w.outBuf, val.Format(timez.DatetimeFormat))
 			case kind.Time:
-				w.fieldPrintFns[i](w.outBuf, val.Format(stringz.TimeFormat))
+				w.fieldPrintFns[i](w.outBuf, val.Format(timez.TimeFormat))
 			case kind.Date:
-				w.fieldPrintFns[i](w.outBuf, val.Format(stringz.DateFormat))
+				w.fieldPrintFns[i](w.outBuf, val.Format(timez.DateFormat))
 			}
 		}
